Using sorted queue for frames
Also reverted frame size change because it would make the lasts frames to be incomplete.

We need to send various frames in the same packet, if we want to improve protocol efficiency.

Or we could try to implement a way to ensure that incomplete frames are completed with silences when recording is stopped so we can send them. Or round or recording side to frames.