Possible fix for remote avatar footstep bug
The bug is that the fake footsteps are played on top of the animation footsteps, most likely because the renderer is returning as "not visible" even though it is visible. Fixed by manually checking whether the avatar object is within the camera view (but only if the renderer is reporting as not visible).