Bucke up, and read the full commit message (...)
The original idea was to fix the preview with this new version of unity,
sadly this is what happened: https://www.youtube.com/watch?v=AbSehcT19u0

List of changes:

1. No more promises in the redux store, having a promise "instanceJS"
was an anti pattern. We now access both unity interface and browser
interface values without a promise.
2. (this may be the most important change) a custom DCLUnityLoader.js is
no longer necessary. We will use the one exposed by unity in the build.
This will lead us to more maintaiable code.
3. The saga that creates the "renderer" now is only one function instead
of three. Specific code for unity and WS modes was moved to other files.
4. MESSAGE_FROM_ENGINE is no longer a saga action. It was pure overhead
for no reason. Messages are now handled directly by the interfaces and
wired at initialization time. (initializeRenderer saga)
5. `gifProcessor` was moved from the DCL global object to the globalThis

Bottom line: the emited code by this newer Unity version is _way_
superior in good practices to the previous code. Global errors nolonger
break unity, and proper stack traces are shown. Not only that, but if an
exception happen, it nolonger freezes the renderer in an unrecoverable
state. We should keep the generated loader.js as pure as possible now
that it doesn't require hacks.